test: keep Workers wire test files within size limits
